Move CoreDataFeedStore to its own file in production.
